Move shared modules to common
