+++ baselines/run_baseline_parallel_fast.py
@@ -7,7 +7,7 @@ from stable_baselines3.common import env_checker
 from stable_baselines3.common.vec_env import DummyVecEnv, SubprocVecEnv
 from stable_baselines3.common.utils import set_random_seed
 from stable_baselines3.common.callbacks import CheckpointCallback, CallbackList
-from tensorboard_callback import TensorboardCallback
+from common.tensorboard_callback import TensorboardCallback
 
 def make_env(rank, env_conf, seed=0):
     """
