+++ common/global_map.py
