+++ common/stream_agent_wrapper.py
@@ -31,9 +31,16 @@ class StreamWrapper(gym.Wrapper):
 
     def step(self, action):
 
-        x_pos = self.emulator.get_memory_value(X_POS_ADDRESS)
-        y_pos = self.emulator.get_memory_value(Y_POS_ADDRESS)
-        map_n = self.emulator.get_memory_value(MAP_N_ADDRESS)
+        if hasattr(self.emulator, "get_memory_value"):
+            x_pos = self.emulator.get_memory_value(X_POS_ADDRESS)
+            y_pos = self.emulator.get_memory_value(Y_POS_ADDRESS)
+            map_n = self.emulator.get_memory_value(MAP_N_ADDRESS)
+        elif hasattr(self.emulator, "memory"):
+            x_pos = self.emulator.memory[X_POS_ADDRESS]
+            y_pos = self.emulator.memory[Y_POS_ADDRESS]
+            map_n = self.emulator.memory[MAP_N_ADDRESS]
+        else:
+            raise AttributeError("Emulator has no supported memory interface")
         self.coord_list.append([x_pos, y_pos, map_n])
 
         if self.steam_step_counter >= self.upload_interval:
