+++ v2/baseline_fast_v2.py
@@ -2,13 +2,13 @@ import sys
 from os.path import exists
 from pathlib import Path
 from red_gym_env_v2 import RedGymEnv
-from stream_agent_wrapper import StreamWrapper
+from common.stream_agent_wrapper import StreamWrapper
 from stable_baselines3 import PPO
 from stable_baselines3.common import env_checker
 from stable_baselines3.common.vec_env import SubprocVecEnv
 from stable_baselines3.common.utils import set_random_seed
 from stable_baselines3.common.callbacks import CheckpointCallback, CallbackList
-from tensorboard_callback import TensorboardCallback
+from common.tensorboard_callback import TensorboardCallback
 
 def make_env(rank, env_conf, seed=0):
     """
