+++ v2/red_gym_env_v2.py
@@ -13,7 +13,7 @@ from einops import repeat
 from gymnasium import Env, spaces
 from pyboy.utils import WindowEvent
 
-from global_map import local_to_global, GLOBAL_MAP_SHAPE
+from common.global_map import local_to_global, GLOBAL_MAP_SHAPE
 
 event_flags_start = 0xD747
 event_flags_end = 0xD87E # expand for SS Anne # old - 0xD7F6
